ci: Improve dependabot configuration
Running dependabot daily results in a pollution of the git history with
dependabot commits. Therefore, it is reconfigured to run monthly.

The commit message prefix is adapted so version bumps are picked up by
release-please.

The automatic merges are squash merges now. This will prevent triangle
merges which were polluting the git history.